删除接口（对传空做判断）

+++ src/main/java/cn/staitech/anno/service/impl/SlideServiceImpl.java
@@ -708,11 +708,13 @@ public class SlideServiceImpl extends ServiceImpl<SlideMapper, Slide> implements
      */
     @Override
     public R deleteProjectImage(ProjectSlideDel projectSlideDel) {
+        if (CollectionUtils.isNotEmpty(projectSlideDel.getSlideIdList())){
         for (Long slideId : projectSlideDel.getSlideIdList()) {
             //切片表删除（物理删）
             slideMapper.deleteProjectImage(slideId);
             //删除文件夹下的图片（物理删除）
             slideMapper.eyeDeleteImage(slideId);
+        }
         }
         return R.ok();
     }
